--- a/index.js
+++ b/index.js
@@ -1,1 +1,7 @@
+var profile = require('./lib/profile');
+
+var evt = profile.event('startup');
+
 module.exports = require('./lib/builder');
+
+evt.done();