--- a/lib/builder.js
+++ b/lib/builder.js
@@ -19,6 +19,8 @@ var Trace = require('./trace');
 
 var getCanonicalName = require('./utils').getCanonicalName;
 
+var profile = require('./profile');
+
 require('rsvp').on('error', function(reason) {
   throw new Error('Unhandled promise rejection.\n' + reason && reason.stack || reason || '' + '\n');
 });
@@ -45,6 +47,7 @@ function Builder(baseURL, cfg) {
 }
 
 Builder.prototype.reset = function(baseLoader) {
+  var evt = profile.event('reset-loader');
   baseLoader = baseLoader || this.loader || System;
 
   var loader = this.loader = new baseLoader.constructor();
@@ -136,6 +139,8 @@ Builder.prototype.reset = function(baseLoader) {
 
   // clear the cache
   this.setCache({});
+
+  evt.done();
 };
 
 Builder.prototype.setCache = function(cacheObj) {
@@ -168,7 +173,7 @@ function executeConfigFile(saveForReset, ignoreBaseURL, source) {
 
 Builder.prototype.loadConfig = function(configFile, saveForReset, ignoreBaseURL) {
   return asp(fs.readFile)(configFile)
-  .then(executeConfigFile.bind(this, saveForReset, ignoreBaseURL));
+  .then(executeConfigFile.bind(this, saveForReset, ignoreBaseURL))
 };
 
 Builder.prototype.loadConfigSync = function(configFile, saveForReset, ignoreBaseURL) {