--- a/lib/compile.js
+++ b/lib/compile.js
@@ -10,6 +10,7 @@ var getTreeModulesPostOrder = require('./trace').getTreeModulesPostOrder;
 var extend = require('./utils').extend;
 var dextend = require('./utils').dextend;
 var getPackage = require('./utils').getPackage;
+var profile = require('./profile');
 
 var compilerMap = {
   'amd': '../compilers/amd',
@@ -94,9 +95,12 @@ function compileLoad(loader, load, compileOpts, cache) {
   if (format == 'defined')
     return Promise.resolve({ source:  compileOpts.systemGlobal + '.register("' + mappedLoad.name + '", [], function() { return { setters: [], execute: function() {} } });\n' });
 
-  if (format in compilerMap)
+  if (format in compilerMap) {
+    var evt = profile.event('compile', load.name);
     return Promise.resolve(require(compilerMap[format]).compile(mappedLoad, compileOpts, loader))
     .then(function(output) {
+      evt.done();
+
       // store compiled output in cache
       cache.loads[load.name] = {
         hash: getCompileHash(load, compileOpts),
@@ -105,12 +109,15 @@ function compileLoad(loader, load, compileOpts, cache) {
 
       return output;
     });
+  }
 
   return Promise.reject(new TypeError('Unknown module format ' + format));
 }
 
 exports.compileTree = compileTree;
 function compileTree(loader, tree, compileOpts, outputOpts, cache) {
+  var evt = profile.event('compile-tree', Object.keys(tree).join(', '));
+
   // sort in graph order, filter modules to actually built loads (excluding conditionals, build: false)
   var ordered = getTreeModulesPostOrder(tree);
   // get entrypoints from graph algorithm
@@ -190,6 +197,7 @@ function compileTree(loader, tree, compileOpts, outputOpts, cache) {
       return addPackageConfig(loader, tree, outputs, entryPoints, compileOpts);
   })
   .then(function(outputs) {
+    evt.done();
     // NB also include all aliases of all entryPoints along with entryPoints
     return {
       outputs: outputs,