--- a/lib/output.js
+++ b/lib/output.js
@@ -3,6 +3,7 @@ var mkdirp = require('mkdirp');
 var fs = require('fs');
 var Promise = require('rsvp').Promise;
 var asp = require('rsvp').denodeify;
+var profile = require('./profile');
 
 var fromFileURL = require('./utils').fromFileURL;
 var toFileURL = require('./utils').toFileURL;
@@ -140,6 +141,7 @@ function inlineSourceMap(source, sourceMap) {
 }
 
 exports.writeOutputs = function(outputs, baseURL, outputOpts) {
+  var evt = profile.event('concatenate');
   var outFile = outputOpts.outFile && path.resolve(outputOpts.outFile);
   var basePath = fromFileURL(baseURL);
   var fileName = path.basename(outFile) || 'output.js';
@@ -153,6 +155,8 @@ exports.writeOutputs = function(outputs, baseURL, outputOpts) {
     output.source = inlineSourceMap(output.source, output.sourceMap);
     output.sourceMap = undefined;
   }
+
+  evt.done();
 
   if (!outputOpts.outFile)
     return Promise.resolve(output);