--- a/lib/trace.js
+++ b/lib/trace.js
@@ -8,6 +8,7 @@ var Graph = require('algorithms/data_structures/graph');
 var depthFirst = require('algorithms/graph').depthFirstSearch;
 var path = require('path');
 var extend = require('./utils').extend;
+var profile = require('./profile');
 
 module.exports = Trace;
 
@@ -27,6 +28,8 @@ Trace.prototype.traceModule = function(moduleName, traceAllConditionals, conditi
   var loader = this.loader;
 
   var self = this;
+
+  var evt = profile.event('deep-trace', moduleName);
   
   return Promise.resolve(loader.normalize(moduleName))
   .then(function(_moduleName) {
@@ -47,6 +50,8 @@ Trace.prototype.traceModule = function(moduleName, traceAllConditionals, conditi
       while ((curMatch = namedRegisterRegEx.exec(thisLoad.source)))
         loads[curMatch[3].substr(1, curMatch[3].length - 2)] = true;
     }
+
+    evt.done();
 
     return {
       moduleName: moduleName,
@@ -72,6 +77,8 @@ Trace.prototype.getLoadRecord = function(canonical) {
 
   if (this.tracing[canonical])
     return this.tracing[canonical];
+
+  var evt = profile.event('trace', canonical);
 
   var self = this;
   var isPackageConditional = canonical.indexOf('#:') != -1;
@@ -329,11 +336,14 @@ Trace.prototype.getLoadRecord = function(canonical) {
         delete load.metadata.builderExecute;
         delete load.metadata.parseTree;
 
+        evt.done();
+
         return load;
       });
     });
   })
   .then(function(load) {
+    evt.cancelIfNotDone();
     return loads[canonical] = load;
   });
 };