--- a/lib/utils.js
+++ b/lib/utils.js
@@ -1,5 +1,6 @@
 var path = require('path');
 var url = require('url');
+var profile = require('./profile');
 
 exports.extend = extend;
 function extend(a, b) {
@@ -64,34 +65,41 @@ function getAlias(loader, canonicalName) {
 
 exports.getCanonicalName = getCanonicalName;
 function getCanonicalName(loader, normalized, isPlugin) {
+  var evt = profile.event('canonicalization', normalized + (isPlugin ? ' (plugin)' : ''));
+
+  var canonical;
+
+  var index, conditionalMatch;
+
   // 1. Boolean conditional
-  var booleanIndex = normalized.lastIndexOf('#?');
-  if (booleanIndex != -1) {
-    var booleanModule = normalized.substr(booleanIndex + 2);
+  if ((index = normalized.lastIndexOf('#?')) != -1) {
+    var booleanModule = normalized.substr(index + 2);
     var negate = booleanModule[0] == '~';
     if (negate)
       booleanModule = booleanModule.substr(1);
-    return getCanonicalName(loader, normalized.substr(0, booleanIndex)) + '#?' + (negate ? '~' : '') + canonicalizeCondition(loader, booleanModule);
+    canonical = getCanonicalName(loader, normalized.substr(0, index)) + '#?' + (negate ? '~' : '') + canonicalizeCondition(loader, booleanModule);
   }
 
   // 2. Plugins
-  var pluginIndex = loader.pluginFirst ? normalized.indexOf('!') : normalized.lastIndexOf('!');
-  if (pluginIndex != -1)
-    return getCanonicalName(loader, normalized.substr(0, pluginIndex), !loader.pluginFirst) + '!' + getCanonicalName(loader, normalized.substr(pluginIndex + 1), loader.pluginFirst);
+  else if ((index = loader.pluginFirst ? normalized.indexOf('!') : normalized.lastIndexOf('!')) != -1) {
+    canonical = getCanonicalName(loader, normalized.substr(0, index), !loader.pluginFirst) + '!' + getCanonicalName(loader, normalized.substr(index + 1), loader.pluginFirst);
+  }
 
   // 3. Package environment map
-  var pkgEnvIndex = normalized.indexOf('#:');
-  if (pkgEnvIndex != -1)
-    return getCanonicalName(loader, normalized.substr(0, pkgEnvIndex), isPlugin) + normalized.substr(pkgEnvIndex);
+  else if ((index = normalized.indexOf('#:')) != -1) {
+    canonical = getCanonicalName(loader, normalized.substr(0, index), isPlugin) + normalized.substr(index);
+  }
 
   // Finally get canonical plain
-  var canonical = getCanonicalNamePlain(loader, normalized, isPlugin);
+  else {
+    canonical = getCanonicalNamePlain(loader, normalized, isPlugin);
+    
+    // 4. Canonicalize conditional interpolation
+    if (conditionalMatch = canonical.match(interpolationRegEx))
+      canonical = getCanonicalNamePlain(loader, normalized, isPlugin).replace(interpolationRegEx, '#{' + canonicalizeCondition(loader, conditionalMatch[0].substr(2, conditionalMatch[0].length - 3)) + '}');
+  }
 
-  // 4. Canonicalize conditional interpolation
-  var conditionalMatch = canonical.match(interpolationRegEx);
-  if (conditionalMatch)
-    return getCanonicalNamePlain(loader, normalized, isPlugin).replace(interpolationRegEx, '#{' + canonicalizeCondition(loader, conditionalMatch[0].substr(2, conditionalMatch[0].length - 3)) + '}');
-
+  evt.done();
   return canonical;
 }