//cc @harrison @jackfranklin 

The first red flag detected in this is that loader reset time is huge, looking into that now.

Sample output:

```
Simple Build
  Total Time               : 1090ms

Startup
  Total Time               : 96ms

Reset Loader
  Total Time               : 277ms

Deep Trace (first.js)
  Total Time               : 627ms

Trace (10 events)
  Cumulative Time          : 1611ms
  Mean                     : 161.1ms
  Std Deviation            : 209.19ms
  Avg within 2σ            : 26.14ms
  Median                   : 32ms
  Max                      : 559ms (component.jsx!jsx.js)
  Min                      : 10ms (some.js!plugin.js)
  Duplicate Events         : 0
  Total Duplicated Time    : 0ms

Compile Tree (first.js, jquery-cdn, second.js, amd.js, third.js, cjs.js, some.js!plugin.js, global.js, jquery.js, text.txt!text-plugin.js, component.jsx!jsx.js)
  Total Time               : 54ms

Compile (10 events)
  Cumulative Time          : 297ms
  Mean                     : 29.7ms
  Std Deviation            : 10.22ms
  Avg within 2σ            : 31.4ms
  Median                   : 31ms
  Max                      : 46ms (third.js)
  Min                      : 16ms (first.js)
  Duplicate Events         : 0
  Total Duplicated Time    : 0ms

Concatenate
  Total Time               : 21ms
```