@guybedford 2015-09-23
When using the build cache, here are the times for the test tree of 10 modules:

```
Cached Tree Build
  Total Time               : 5ms

Deep Trace (first.js)
  Total Time               : 1ms

Compile Tree (first.js, second.js, amd.js, component.jsx!jsx.js, third.js, cjs.js, global.js, some.js!plugin.js, text.txt!text-plugin.js, jquery.js, jquery-cdn)
  Total Time               : 1ms

Concatenate
  Total Time               : 0ms
```

This shows that the cache is working pretty well.